Updating to latest main

+++ src/main/java/com/capgemini/flightbookingsystem/controllers/UserController.java
@@ -37,7 +37,7 @@ public class UserController {
 	}
 
 	@GetMapping("{id}")
-	public ResponseEntity<User> getUser(@PathVariable Long id) {
+	public ResponseEntity<User> getUser(@PathVariable Integer id) {
 		User user = userService.getUserById(id);
 		return ResponseEntity.status(HttpStatus.OK).body(user);
 	}
@@ -50,19 +50,19 @@ public class UserController {
 	}
 
 	@PutMapping("/{id}")
-	public ResponseEntity<User> updateUserRegister(@PathVariable Long id, @RequestBody User newUser) {
+	public ResponseEntity<User> updateUserRegister(@PathVariable Integer id, @RequestBody User newUser) {
 		User updated = userService.putUser(id, newUser);
 		return ResponseEntity.status(HttpStatus.OK).body(updated);
 	}
 
 	@PatchMapping("/{id}")
-	public ResponseEntity<User> patchUser(@PathVariable Long id, @RequestBody User patch) {
+	public ResponseEntity<User> patchUser(@PathVariable Integer id, @RequestBody User patch) {
 		User updated = userService.patchUser(id, patch);
 		return ResponseEntity.status(HttpStatus.OK).body(updated);
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+	public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
 		userService.deleteUser(id);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 	}
